% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blacklist_Groups.R
\name{Blacklist_Groups}
\alias{Blacklist_Groups}
\title{Blacklist Groups}
\usage{
Blacklist_Groups(data, groups, rhop, centroid_flag, log_file_name)
}
\arguments{
\item{data}{Processed data from Clean_Up_Input (or Remove_Cell_Cycle)}

\item{groups}{Processed groups file from Clean_Up_Input}

\item{rhop}{x in mean+x*SD to determine upper cutoff for correlation in the blacklist. Default is 1.}

\item{centroid_flag}{use centroids if data is sparse}

\item{log_file_name}{used for saving run notes to log file}
}
\value{
newMedoids - new medoids data.frame for the new combined blacklisted clusters.

newGroups - new groups file containing cluster assignment based on new combined blacklisted clusters.
}
\description{
This function is used for the calculation of blacklisted medoids and groups. It calculates medoids. It performs medoid correlations. It creates the binary correlation table between medoids, otherwise known as the blacklist. It makes a blacklist heatmap. It makes a blacklist heatmap. It makes new medoids based on blacklisted combined clusters. It makes a new groups file based on the blacklisted combined clusters.
}
\keyword{blacklist}
\keyword{correlation}
