% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Up_Input.R
\name{Clean_Up_Input}
\alias{Clean_Up_Input}
\title{Clean Up Input}
\usage{
Clean_Up_Input(rawData, groups, rowClusters = NULL, log_file_name)
}
\arguments{
\item{rawData}{ICGS expression or counts file (in ICGS expression file format).}

\item{groups}{ICGS groups file.}

\item{rowClusters}{Optional vector containing row cluster information (used by Recluster). Default is NULL.}

\item{log_file_name}{used for saving run notes to log file}
}
\value{
processed - data.frame of genes by samples with a row of cell clusters (column_clusters-flat) and a column of gene clusters (row_clusters-flat) when available.

groups - groups file with cell names matching the expression file.
}
\description{
This function cleans up the raw data and groups files. It replaces the column clusters in the expression file with the clusters in the groups file. It ensures that all expression files have a column_clusters-flat and row_clusters-flat (even when data not provided it adds NAs). It uses Renumber as necessary to avoid non-consecutive cluster identifications.
}
\keyword{clean}
