% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hopach_and_Heatmap.R
\name{Hopach_and_Heatmap}
\alias{Hopach_and_Heatmap}
\title{Hopach and Heatmap}
\usage{
Hopach_and_Heatmap(partialData, fullData, filename, groups)
}
\arguments{
\item{partialData}{Subset of data, either doublets or nondoublets called by DeconRNASeq.}

\item{fullData}{Entire processed data from Clean_Up_Data.}

\item{filename}{Unique filename to be incorporated into the names of outputs from the functions.}

\item{groups}{Processed groups file from Clean_Up_Input.}
}
\value{
partialGroups - new groups file for the partialData based on hopach clustering.
}
\description{
This function performs hopach clustering and creates a heatmap and groups file.
}
\keyword{HOPACH}
\keyword{heatmap}
