% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Is_A_Doublet.R
\name{Is_A_Doublet}
\alias{Is_A_Doublet}
\title{Is A Doublet}
\usage{
Is_A_Doublet(data, newMedoids, groups, synthProfiles, log_file_name)
}
\arguments{
\item{data}{Processed data from Clean_Up_Input (or Remove_Cell_Cycle).}

\item{newMedoids}{New combined medoids from Blacklist_groups.}

\item{groups}{Processed groups file from Clean_Up_Input.}

\item{synthProfiles}{Average profiles of synthetic doublets from Synthetic_Doublets.}

\item{log_file_name}{used for saving run notes to log file}
}
\value{
isADoublet - data.frame with each cell as a row and whether it is called a doublet by deconvolution analysis.

resultsreadable - data.frame with results of deconvolution analysis (cell by cluster) in percentages.
}
\description{
This function uses deconvolution analysis (DeconRNASeq) to evaluate each cell for equal contribution from blacklisted clusters.
}
\keyword{decon}
\keyword{deconvolution}
\keyword{doublet}
