% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Doublet_Decon.R
\name{Main_Doublet_Decon}
\alias{Main_Doublet_Decon}
\title{Main DoubletDecon}
\usage{
Main_Doublet_Decon(rawDataFile, groupsFile, filename, removeCC = FALSE,
  species = "mmu", corrCutoff = 1, write = TRUE, recluster = "none",
  isADoublet = 1)
}
\arguments{
\item{rawDataFile}{Name of file containing expression data (gene by cell)}

\item{groupsFile}{Name of file containing group assignments (3 column: cell, group(numeric), group(numeric or character))}

\item{filename}{Unique filename to be incorporated into the names of outputs from the functions}

\item{removeCC}{Remove cell cycle gene cluster by KEGG enrichment. Default is FALSE.}

\item{species}{Species as scientific species name, KEGG ID, three letter    species abbreviation, or NCBI ID. Default is "mmu".}

\item{corrCutoff}{x in mean*x*SD to determine upper cutoff for correlation in the blacklist. Default is 1.}

\item{write}{Write output files as .txt files. Default is TRUE.}

\item{recluster}{Recluster deconvolution classified doublets and non-doublets deperately using hopach. Default is FALSE.}

\item{isADoublet}{Use Decon Correlation method (1) or Two Percentages method (2). Default is 2.}
}
\value{
data_processed - new expression file (cleaned).

groups_processed = new groups file (cleaned).

PMF_results_1 = pseudo marker finder t statistics (gene by cluster).

PMF_results_2 = pseudo marker finder gene assignments (which cluster has the highest t-stat for each gene).

DRS_doublet_table = each cell and whether it is called a doublet by deconvolution analysis.

DRS_results = results of deconvolution analysis (cell by cluster) in percentages.

Decon_called_freq = percentage of doublets called in each cluster by deconvolution analysis.

Final_doublets_groups = new groups file containing only doublets.

Final_nondoublets_groups = new groups file containing only non doublets.
}
\description{
This is the main function. This function identifies clusters of doublets with a combination of deconvolution analysis and unique gene expression and individual doublet cells with deconvolution analysis.
}
\keyword{decon}
\keyword{doublet}
\keyword{main}
