% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudo_Marker_Finder.R
\name{Pseudo_Marker_Finder}
\alias{Pseudo_Marker_Finder}
\title{Pseudo MarkerFinder}
\usage{
Pseudo_Marker_Finder(groups, data, full_data2)
}
\arguments{
\item{groups}{Processed groups file from Clean_Up_Input.}

\item{data}{Processed data from Clean_Up_Input (or Remove_Cell_Cycle).}

\item{full_data2}{Variable for passing the full gene expression matrix.}
}
\value{
hallmarkTable - pseudo marker finder t statistics (gene by cluster).

hallmarkTable2 - pseudo marker finder gene assignments (which cluster has the highest t-stat for each gene).
}
\description{
This function uses t-tests to look for unique gene expression in each cluster.
}
\keyword{Finder}
\keyword{Marker}
\keyword{ttest}
