% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renumber.R
\name{Renumber}
\alias{Renumber}
\title{Renumber}
\usage{
Renumber(x)
}
\arguments{
\item{x}{Vector to be renumbered}
}
\value{
y - renumbered vector
}
\description{
This function renumbers a vector to avoid non-consecutive cluster lables generated from cell cycle removal, hopach clustering, or non-standard raw data input.
}
\keyword{renumber}
