% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Pre_Process.R
\name{Seurat_Pre_Process}
\alias{Seurat_Pre_Process}
\title{Seurat Pre Process}
\usage{
Seurat_Pre_Process(expressionFile, genesFile, clustersFile)
}
\arguments{
\item{expressionFile}{Normalized expression file from Seurat}

\item{genesFile}{Gene list from Seurat}

\item{clustersFile}{Cluster list from Seurat}
}
\value{
newExpressionFile - Seurat expression file in ICGS format (ICGS genes)

newFullExpressionFile - Seurat expression file in ICGS format (all genes)

newGroupsFile - Groups file ICGS format
}
\description{
This function has been replaced with Improved_Seurat_Pre_Process for Seurat Version 3.
}
\keyword{Seurat}
