% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthetic_Doublets.R
\name{Synthetic_Doublets}
\alias{Synthetic_Doublets}
\title{Synthetic Doublets}
\usage{
Synthetic_Doublets(data, groups, groupsMedoids, newMedoids)
}
\arguments{
\item{data}{Processed data from Clean_Up_Input (or Remove_Cell_Cycle).}

\item{groups}{Processed groups file from Clean_Up_Input.}

\item{groupsMedoids}{New groups file based on blacklisted clusters from Blacklist_Groups}

\item{newMedoids}{New combined medoids from Blacklist_groups.}
}
\value{
averagesAverages - average deconvolution profiles for each combination of cell types.
}
\description{
This function creates synthetic doublets by averaging gene expression profiles from each combination of clusters to generate deconvolution profiles for each type of doublet.
}
\keyword{synthetic}
