% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthetic_Doublets_test.R
\name{Synthetic_Doublets_test}
\alias{Synthetic_Doublets_test}
\title{Synthetic Doublets (for testing)}
\usage{
Synthetic_Doublets_test(rawDataFile, groupsFile, filename, removeCC = FALSE,
  species = "mmu", rhop = 1.15, numDubPercent = 0.15, full_PMF = FALSE,
  fullDataFile = NULL)
}
\arguments{
\item{rawDataFile}{Name of file containing expression data (gene by cell)}

\item{groupsFile}{Name of file containing group assignments (3 column: cell, group(numeric), group(numeric or character))}

\item{filename}{Unique filename to be incorporated into the names of outputs from the functions}

\item{removeCC}{Remove cell cycle gene cluster by KEGG enrichment. Default is FALSE.}

\item{species}{Species as scientific species name, KEGG ID, three letter    species abbreviation, or NCBI ID. Default is "mmu".}

\item{rhop}{x in mean*x*SD to determine upper cutoff for correlation in the blacklist. Default is 1.}

\item{numDubPercent}{number of cells to be created as doublets based on percentage of total cells. Default is 0.15.}
}
\description{
This function creates synthetic doublets by averaging gene expression profiles from each combination of clusters to generate deconvolution profiles for each type of doublet. Doublets should be incorporated into data with cellHarmony prior to testing.
}
\keyword{synthetic}
