% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudo_Marker_Finder.R
\name{Pseudo_Marker_Finder}
\alias{Pseudo_Marker_Finder}
\title{Pseudo MarkerFinder}
\usage{
Pseudo_Marker_Finder(groups, redu_data2, full_data2, min_uniq = 4,
  log_file_name = log_file_name)
}
\arguments{
\item{groups}{Processed groups file from Clean_Up_Input.}

\item{redu_data2}{Processed data from Clean_Up_Input (or Remove_Cell_Cycle) path, automatically written.}

\item{full_data2}{cleaned full expression matrix from Clean_Up_Input.}

\item{min_uniq}{minimum number of unique genes required for a cluster to be rescued}

\item{log_file_name}{used for saving run notes to log file}

\item{nCores}{number of cores to be used during rescue step. Default is -1 for automatically detected.}
}
\value{
new_table - non-doublet clusters, as determined by the "Remove" and "Rescue" steps.
}
\description{
This function uses ANOVA to look for unique gene expression in each possible doublet cluster.
}
\keyword{ANOVA}
\keyword{Finder}
\keyword{Marker}
