% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Pre_Process.R
\name{Seurat_Pre_Process}
\alias{Seurat_Pre_Process}
\title{Seurat Pre Process}
\usage{
Seurat_Pre_Process(expressionFile, genesFile, clustersFile)
}
\arguments{
\item{expressionFile}{Normalized expression file from Seurat}

\item{genesFile}{Gene list from Seurat}

\item{clustersFile}{Cluster list from Seurat}
}
\value{
newExpressionFile - Seurat expression file in ICGS format

newGroupsFile - Groups file ICGS format
}
\description{
This function uses t-tests to look for unique gene expression in each cluster.
}
\keyword{Seurat}
